requirements.txt
Django==2.2.1
django-extensions==2.2.3
djangorestframework==3.10.3
psycopg2-binary==2.8.2
python-decouple==3.1
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
